asio_http_server::file_server byte range requests
...of course there is support for all conditional request's preconditions too

and about the HTTP/1.0 , it's a collateral effect of being able to completely separate HTTP producers (that might or might not have native streaming... or might not even be HTTP!) and HTTP consumers (who cannot rely on the HTTP version to guess available features... it may not even have a HTTP version). But the implementation (for the producer and for the consumer, not for the gluing layer) is correct. If an HTTP 1.1 request have been made, the streaming interface (chunked entities) would be used instead.

Overlapping will be merged, of course. But how else would you be supposed to know the validating, splitting, file manipulation, too name a few, are correct?

Implementation is almost in shape, validating every input and giving an answer that will do the right thing. Of course interface is more worth, that's why the interface is very reusable.